requirements.txt
PyYAML
requests
python-ldap
